Add RenderText to rive_native
This lets the editor (and users of rive_native) draw text procedurally a little more easily directly through the runtime's RawText feature. This makes it easier to use all the high level features of our Text object in a procedural way, which is helpful for the editor to use on the Stage. Based on discussions with @alxgibsn where we found that there wasn't a good way to draw text with ellipses on the stage.

You could (and we did) previously do this by using helpers like DisplayText which would do all the styling and glyph rendering in Dart. This will be more performant for cases where you don't need control of each individual glyph from Dart and just want to draw a blob of styled text. It will incur much less FFI overhead and should be used by the editor when possible for text handles/helpers (Rive Text should still go through the usual path as modifiers need per glyph control, Raw/RenderText does not work with modifiers).

Diffs=
de0026bd59 Add RenderText to rive_native (#9146)